Version 3.6

+++ client/src/app/components/navigation-usuario/navigation-usuario.component.ts
@@ -4,7 +4,9 @@ import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 
 //importamos servicio
-import { ProductosService } from '../../services/productos.service'
+import { ProductosService } from '../../services/productos.service';
+
+import { Cliente } from 'src/app/models/registroCliente';
 
 @Component({
   selector: 'app-navigation-usuario',
@@ -16,12 +18,27 @@ export class NavigationUsuarioComponent implements OnInit {
   constructor(private router:Router, private service:ProductosService) { }
 
   ngOnInit(): void {
+    
   }
 
   cerrarSecion(){
-    this.service.logoutLS();
-    this.service.deleteCarritoLS();
-    this.router.navigate(['/login']); //si no esta logueado me redirije a login para que meta sus datos
+    let d_json=this.service.getClienteLS();
+    if(d_json){
+      //guardo id de Cliente
+      let cliente:Cliente=d_json;
+      let fecha=new Date();
+      let fechaa=fecha.getDate()+'-'+(fecha.getMonth()+1)+'-'+fecha.getFullYear()+' : '+fecha.getHours()+':'+fecha.getMinutes();
+      //guardo en bitacora
+      //guardo en bitacora
+      this.service.saveBitacora(cliente.correo,"Cerro Sesion",fechaa).subscribe(
+        res=>{
+          this.service.logoutLS();
+          this.service.deleteCarritoLS();
+          this.router.navigate(['/login']); //si no esta logueado me redirije a login para que meta sus datos
+        },
+        err=>console.error(err)
+      ); 
+    }    
   }
 
   perfil(){
